Update flags.class.php
It's more sane now, if only marginally. Calling a method without a value
returns the value - this allows flags to be used as callbacks with no
wrapper function. Also, flags now store their existence even when false,
so iteration actually has a use now. In order to forget a flag, set it
to null.